refactor firebase events to analytics file
