+++ app/page.js
@@ -5,7 +5,7 @@ import logo from "/public/murphylogo.png";
 import Link from "next/link";
 import Button from "@/components/general/Button";
 import { usePageAnalytics } from "@/app/utils/useAnalytics";
-import { logButtonEvent, logLoadingTime } from "@/app/firebaseConfig";
+import { logButtonEvent, logLoadingTime } from "@/app/utils/analytics";
 import { useEffect } from "react";
 export default function Home() {
   usePageAnalytics("/");
